mcp_spartacus/requirements.txt
fastapi
uvicorn
langchain==0.3.27
langchain-core==0.3.72
langchain-gemini
langchain-google-genai
langchain-text-splitters==0.3.9
langgraph==0.6.1
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langsmith==0.4.8
google-generativeai
sqlalchemy<2.0.0
psycopg2-binary<2.10.0
django<4.0.0
python-dotenv
